% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_skew_stats.R
\name{compute_skew_stats}
\alias{compute_skew_stats}
\title{Compute alternative boxplot statistics}
\usage{
compute_skew_stats(x, method = "tukey", k = 1.5)
}
\arguments{
\item{x}{A numeric vector}

\item{method}{Method name ("tukey", "kimber", "hubert", etc.)}

\item{k}{Tuning parameter (default 1.5)}
}
\value{
A list of boxplot stats: ymin, lower, middle, upper, ymax, outliers
}
\description{
Compute alternative boxplot statistics
}
