% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_skewboxplot.R
\docType{data}
\name{StatSkewBoxplot}
\alias{StatSkewBoxplot}
\title{Statistical Transformation for Skew Boxplots}
\format{
A ggproto object
}
\description{
\code{StatSkewBoxplot} is a ggproto object that computes statistics necessary
to create skew boxplots, including quantiles, skewness measures, and whiskers.
}
\keyword{internal}
